LayoutTests/editing/pasteboard/dataTransfer-setData-getData-expected.txt

Tests correct behavior of event.dataTransfer.setData/getData for 'URL', 'text/uri-list' and 'text/plain'

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


--- Test set/get 'URL':
PASS getDataResultType is "string"
PASS getDataLines.length is 1
PASS getDataLines[0].replace(/\/$/, '') is "http://test.com"
--- Test set/get 'text/uri-list':
PASS getDataResultType is "string"
PASS getDataLines.length is 2
PASS getDataLines[0].replace(/\/$/, '') is "http://test.com"
PASS getDataLines[1].replace(/\/$/, '') is "http://check.com"
--- Test set 'text/uri-list', get 'URL':
PASS getDataResultType is "string"
PASS getDataLines.length is 1
PASS getDataLines[0].replace(/\/$/, '') is "http://test.com"
--- Test set 'text/uri-list', get 'URL', using only '\n':
PASS getDataResultType is "string"
PASS getDataLines.length is 1
PASS getDataLines[0].replace(/\/$/, '') is "http://test.com"
--- Test set/get 'text/uri-list' with comments:
PASS getDataResultType is "string"
PASS getDataLines.length is 2
PASS getDataLines[0].replace(/\/$/, '') is "http://test.com"
PASS getDataLines[1].replace(/\/$/, '') is "http://check.com"
--- Test set/get 'text/plain':
PASS getDataResultType is "string"
PASS getDataLines.length is 1
PASS getDataLines[0].replace(/\/$/, '') is "Lorem ipsum dolor sit amet."
PASS successfullyParsed is true

TEST COMPLETE
